增加雨量相关逻辑层处理，还缺视图层相关代码。

+++ CJSW-WebMVC/Models/Record.cs
@@ -11,6 +11,6 @@ namespace CJSW_WebMVC.Models
     public class Record
     {
         public Station station;
-        public List<Nullable<double>> datas;
+        public List<Nullable<decimal>> datas;
     }
 }
